v0.0.12. Modify the test to check the manager of rod components

+++ src/core/aliasdata.h
@@ -15,7 +15,7 @@ namespace QRS::Core
 
 using DataValueType = double;
 using DataKeyType = double;
-using DataIDType = quint32;
+using DataIDType = quint64;
 
 }
